requirements.txt
autogen
requests
pytest
python-dotenv
openai
autogen-agentchat
autogen-ext[models,magentic-one,openai,azure]
FastAPI
uvicorn
pydantic
azure.identity
azure.monitor.query
aiohttp